refactor: split Import with AI tabs into separate pages
- Move AI URL, Text, YouTube, and TikTok tabs from openai.vue to separate pages
- Add video icon for TikTok/YouTube menu items
- Update Create menu to show all import options directly
- Remove openai.vue (tabs now as standalone pages)
- Remove docker/docker-compose.yml (keep only root compose.yaml)
