--- docker/docker-compose.yml
@@ -1,38 +1,0 @@
-services:
-  mealie:
-    container_name: mealie
-    image: mealie:dev
-    build:
-      context: ../
-      target: production
-      dockerfile: ./docker/Dockerfile
-    restart: always
-    volumes:
-      - mealie-data:/app/data/
-    ports:
-      - 9091:9000
-    environment:
-      ALLOW_SIGNUP: "false"
-      LOG_LEVEL: "DEBUG"
-
-      DB_ENGINE: sqlite # Optional: 'sqlite', 'postgres'
-      # =====================================
-      # Postgres Config
-      POSTGRES_USER: mealie
-      POSTGRES_PASSWORD: mealie
-      POSTGRES_SERVER: postgres
-      POSTGRES_PORT: 5432
-      POSTGRES_DB: mealie
-      # =====================================
-      # Email Configuration
-      # SMTP_HOST=
-      # SMTP_PORT=587
-      # SMTP_FROM_NAME=Mealie
-      # SMTP_AUTH_STRATEGY=TLS # Options: 'TLS', 'SSL', 'NONE'
-      # SMTP_FROM_EMAIL=
-      # SMTP_USER=
-      # SMTP_PASSWORD=
-
-volumes:
-  mealie-data:
-    driver: local
