+++ frontend/components/Layout/DefaultLayout.vue
@@ -208,9 +208,27 @@ export default defineNuxtComponent({
       {
         insertDivider: false,
         icon: $globals.icons.robot,
-        title: i18n.t("recipe.import-with-ai"),
-        subtitle: i18n.t("recipe.import-with-ai-subtitle"),
-        to: `/g/${groupSlug.value}/r/create/openai`,
+        title: i18n.t("recipe.ai-tab-text"),
+        subtitle: i18n.t("recipe.ai-text-description"),
+        to: `/g/${groupSlug.value}/r/create/text`,
+        restricted: true,
+        hide: !$appInfo.enableOpenai,
+      },
+      {
+        insertDivider: false,
+        icon: $globals.icons.video,
+        title: i18n.t("recipe.ai-tab-youtube"),
+        subtitle: i18n.t("recipe.ai-youtube-description"),
+        to: `/g/${groupSlug.value}/r/create/youtube`,
+        restricted: true,
+        hide: !$appInfo.enableOpenai,
+      },
+      {
+        insertDivider: false,
+        icon: $globals.icons.video,
+        title: i18n.t("recipe.ai-tab-tiktok"),
+        subtitle: i18n.t("recipe.ai-tiktok-description"),
+        to: `/g/${groupSlug.value}/r/create/tiktok`,
         restricted: true,
         hide: !$appInfo.enableOpenai,
       },
