+++ frontend/lib/icons/icons.ts
@@ -147,6 +147,7 @@ import {
   mdiTools,
   mdiTranslate,
   mdiUndo,
+  mdiVideo,
   mdiViewDashboard,
   mdiViewModule,
   mdiWeatherNight,
@@ -280,6 +281,7 @@ export const icons = {
   upload: mdiCloudUpload,
   viewDashboard: mdiViewDashboard,
   viewModule: mdiViewModule,
+  video: mdiVideo,
   weatherNight: mdiWeatherNight,
   weatherSunny: mdiWeatherSunny,
   webhook: mdiWebhook,
